\name{remind_me}
\alias{remind_me}
\title{Reminder of important things}
\usage{
remind_me()
}
\description{
Prints everything important that the author of this package needs to do.
}
\examples{
remind_me()

#Output
Birthdays:
 Zoli 4.2.
 Fran 6.2.
 Vukasin 26.2

Shopping list:
 Rakija
 Ajvar
 Kobasica

}
